Refactor: rds, redis버전으로 구분지어서 관리하도록 변경
