+++ gathering-service/src/main/java/com/sparta/moim/gathering/gathering/application/service/rds/GatheringService.java
@@ -1,4 +1,4 @@
-package com.sparta.moim.gathering.gathering.application.service;
+package com.sparta.moim.gathering.gathering.application.service.rds;
 
 import com.sparta.moim.gathering.gathering.application.dto.command.CreateGatheringCommand;
 import com.sparta.moim.gathering.gathering.application.dto.command.DeleteGatheringCommand;
